gpiolib: use dev_gpiod_get_index device node argument throughout
dev_gpiod_get_index takes both a consumer device and a consumer device
tree node. The consumer device tree node should be used for all lookups,
but it was used only for one of two lookups so far.

This would break using the function on subnodes of a device, so remedy
that.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240809142405.315244-9-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>