ci: container: update to Debian Bookworm with GCC v14.2.0
We have been using the same distro/toolchain for more than a year now,
so let's update them to their newest versions.

pip3 now warns that we are not in a venv. Acknowledge this and ignore
it as we want labgrid to be available to pytest in the container without
having to source a virtualenv first.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240813070201.1069847-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>