ci: container: add kvx-elf- toolchain
ARCH=kvx build has been broken in the past, because it was the only
barebox platform that wasn't build tested.

There are prebuilt toolchains of acceptable footprint (64M compressed),
so let's integrate them into the image, so we can add it to build test.

There's no prebuilt Qemu unfortunately, so actually testing the result
will have to wait.

Acked-by: Yann Sionneau <ysionneau@kalrayinc.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240813070201.1069847-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>