ARM: i.MX: drop explicit reservation of scratch area
The scratch area is just some unused space between top of stack and end
of early memory / start of OP-TEE. There's nothing i.MX specific about
it, except that we use it to store the bootrom log.

The initcall reserving the scratch area was added to avoid barebox
proper overwriting it later on during boot.

Since commit 59289e3d8cc3 ("memory: add support for requesting barebox
area as a whole"), the scratch area is reserved as part of the barebox
memory region, so it's unnecessary to keep around SoC-specific code
for its reservation, so let's just drop it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240814093824.71656-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>