include: linux/math.h: promote DIV_ROUND_UP_ULL args to 64-bit
DIV_ROUND_DOWN_ULL promotes division arguments to 64-bit, but
DIV_ROUND_UP_ULL did so only for the division, not the addition for the
round up. This would lead to a wrong result when the 32-bit addition
wraps around.

Linux has an explicit cast to fix this, so do likewise in barebox.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240815065646.1156788-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>