MIPS: dts: add /chosen/stdout-path where it is missing
A number of MIPS boards don't have a stdout-path set and instead rely on
CONFIG_CONSOLE_ACTIVATE_ALL/FIRST to enable a console.

This can get problematic if more consoles are added in future, or
CONFIG_CONSOLE_ACTIVATE_NONE is to be used. So on systems that currently
have a single console, let's have /chosen/stdout-path point at it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240822115701.1184099-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>