test: conftest.py: move top-level
We have a top level pytest.ini to support running pytest --lg-env=$some_env
without extra arguments. This works for running tests nornally, but not if
extra options like --qemu should be used, because the conftest.py won't
be loaded early enough.

The easy fix is moving the conftest.py a level higher as described in:

  https://github.com/pytest-dev/pytest/issues/5822#issuecomment-697331920

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240822115730.1184500-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>