openrisc: dts: describe all 4 UARTs
Starting with QEMU v7.1.0, the openrisc_sim machine features 4 serial
ports. This was introduced with QEMU commit 777784bda468 ("hw/openrisc:
support 4 serial ports in or1ksim"), which also made the last console
(serial@90000300) the default.

This is still the case as of v9.1.0-rc3, but I submitted a patch to
restore serial@90000000 as default. Until that's merged, let's just
activate all consoles after describing them in the device tree.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240822165633.3726922-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>