test: directly exit when barebox state transition fails
For trivial errors like a QEMU -bios or -kernel option not pointing at
an existing file, the pytest output is extremely confusing. It prints a
stack trace for every single test after QMPError("Received empty response")
is raised and the stderr text explaining that the file is missing is
easily overlooked.

Improve this by just existing right away once a strategy transition
fails and print the standard error if any as part of the exit message.

This looks a bit cumbersome, because the capsys fixture can't be used
with a session-scoped fixture, see the still open pytest issue[1] for
details.

[1]: https://github.com/pytest-dev/pytest/issues/2704

Suggested-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240822191714.3768641-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>