fbconsole: support fonts wider than 16 pixels
We have a nested loop that iterates over all pixels in a font, but it
only looks at the first 8 bits. This worked so far, because our biggest
font was VGA8x16. In the follow-up commit, a 16x32 font will be added,
so we need to prepare the code to consider the whole width of a font.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240824200406.2160679-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>