include: linux/iopoll: fix uninitialized warning
In gcc 13.3 there is a warning that start may be used unitialized:

  include/linux/iopoll.h:42:21: warning: ‘start’ may be used uninitialized [-Wmaybe-uninitialized]
     42 |                     is_timeout(start, ((timeout_us) * USECOND))) { \
        |                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The warning is bogus since before usage of start !IN_PBL and
timeout_us != 0 are checked for the case where start is used, but in
this case it is also always initialized to get_time_ns().
Initialize it to zero to silence the warning.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
Link: https://lore.barebox.org/20240903093712.453901-1-r.czerwinski@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>