ARM: rpi: return correct error code when failing to read revision
When rpi_get_dcfg() returned an error, ret wasn't set and continued to
hold the value of the hardware revision, which would be used as return
value for the initcall.

rpi_get_dcfg() already prints an error message with the HW revision when
returning an error pointer, so passing along the extra error code
results in no loss of information.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240911081747.1523613-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>