efi: payload: image use new CONFIG_ARM64 symbol
CONFIG_ARM64 was added as an alternative to CONFIG_CPU_V8 as encountering
the latter in generic code is confusing for users not familiar with ARM.

EFI payload support was one such generic place, so let's replace the
CONFIG_CPU_v8 instance there with the new symbol.

The if clause is moved to keep the order elsewhere in the file, where
x86 checks are done before those for ARM. Both can't be active at the
same time, so no functional change expected.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240920094900.1838301-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>