commands: splash: honour line_length
the splash command has an option for setting a background colour. This
doesn't always fill the whole screen as expected.

struct fb_info has a line_length field that specifies the length in
bytes of a single line. That length may be bigger than the actually
visible area as specified by xres. In this case filling the area as
calculated by xres*yres fills the invisible pixels between xres and
line_length, but it doesn't fill the whole visible area. Fix this by
iterating over the lines, filling the visible area of each line.

Link: https://lore.barebox.org/20240925-graphics-v1-2-f1fe5d2cceec@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>