spi: spi-gpio: switch to new gpio binding
The old deprecated device tree binding for the spi-gpio driver uses
"gpio-sck", "gpio-mosi" and "gpio-miso" to specify the GPIOs. Switch to
the new binding which uses the standard GPIO property names.

The old binding is still used in some device trees, but none of the SoCs
using them is actually supported in barebox, so do not bother to keep a
fallback to the old binding. The one in-tree user of the old binding is
converted in this patch.

Link: https://lore.barebox.org/20240926-spi-gpio-v2-3-3bf7b511c8c0@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>