video: i.MX ipuv3: switch to upstream hdmi driver
The Linux dw-hdmi driver is based on a driver which originally was
written specifically for the i.MX and was converted later to a generic
dw-hdmi driver. The barebox driver is still based on the original Linux
driver. We just merged the dw-hdmi driver. With this patch we now get
rid of the i.MX specific driver and use the generic one.

Link: https://lore.barebox.org/20240927-vop2-v2-9-dc8dcfc651d5@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>