video: Rockchip: Add VOP2 driver
This adds support for the Rockchip VOP2 video core found on the Rockchip
SoCs RK3566, RK3568 and RK3588. The code is based on the Linux driver
and has been heavily stripped down for barebox. Support for the cluster
windows has been removed, also support for YUV modes has been dropped.

Tested-by: Ahmad Fatoum <a.fatoum@pengutronix.de> # rk3566 HDMI
Link: https://lore.barebox.org/20240927-vop2-v2-15-dc8dcfc651d5@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>