test: arm: qemu-sabrelite: Use qemu specific device tree
The sabrelite board has some GPIOs registered as buttons. These GPIOs
are not properly emulated in QEMU and read as low. The buttons are
active low as well, so in barebox we detect all buttons as pressed. One
of these buttons is registered as KEY_POWER. Depending on the
initialization order it can happen that barebox tries to power down the
board during startup. Avoid this pitfall by passing a device tree to
QEMU which has the gpio-keys driver disabled.

Link: https://lore.barebox.org/20241001085208.522095-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>