ARM: Wolfvision PF5: of_probe only once
When overlays have been applied we have to call of_clk_init() and
of_probe() once to have the changes from the overlays take effect.  We
only have to do it once though after all overlays have been applied, not
after each overlay.

Doing it after each overlay has the undesired side effect that the VOP2
will be probed right after the io-expander board has been detected. At
that time the display overlay is not yet included, so the VOP2 probes
without the display and it won't be available later.

Link: https://lore.barebox.org/20241001104738.1069237-2-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>