test: pytest: introduce pytest for network, test tftp notfound
barebox already has a few pytest integration tests but none of them are
networking related.

This patch introduces a test that asserts on bringing up at least one
DHCP interface and, on any of these interfaces, a successful ICMP-ping
from the DUT to the test pytest-host as well as a basic "File not Found"
conversation between the DUT and the test pytest-host.

As the test requires a network connection to the barebox device, without
the "network" feature the test is marked xfail. For the case of a qemu
device we also assume DHCP to set serverip correctly. In any other case,
it is set via console to that IP of the pytest-host that has the route
to the barebox devices IP.

Signed-off-by: Jonas Rebmann <mail@schlaraffenlan.de>
Tested-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20241008-test_tftp-v2-1-509b8e861702@schlaraffenlan.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>