ARM64: asm: mark assembly entry points as functions
ENTRY_PROC_END() is used as counterpart to ENTRY_PROC() to delimit entry
points written in assembly. That macro expands to the END() macro though,
which doesn't explicitly mark the entry point as being a function.

This can lead to objdump -d on the PBL ELF file to not disassemble
the wntry point and instead displaying only a hexdump.

Fix this by using ENDPROC which sets the function type correctly.
This only has effect on the ELF and introduces no functional change for
the objcopied image.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20241009055731.4115395-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>