sandbox: defconfig: activate CONFIG_CONSOLE_ACTIVATE_ALL
The default CONSOLE_ACTIVATE value set in common/Kconfig is
CONSOLE_ACTIVATE_ALL_FALLBACK.

Which adds an error when there's no explicit activation for a specific
console, that is important for real hardware as it may write barebox log
to external devices that don't expect it.

But in case of sandbox mode there's no explicit activation and the error
is always there in spite of number of consoles and there are no external
devices to worry about.

Fixes: 219b2341dc16 ("console: add new CONSOLE_ACTIVATE_ALL_FALLBACK option")
Signed-off-by: Mohamed Sawy <mohamedsawy1022@gmail.com>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20241011201818.42338-1-mohamedsawy1022@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>