sandbox: lds: fix relocation in read-only section warning
To address an "ELF has a LOAD segment with RWX permissions" warning
during the sandbox build, the RO_DATA_SECTION was marked READONLY.

At least with Debian GCC 12.2.0-14, this just brought us a different
warning:

  /usr/bin/ld: common/memory.o: warning: relocation in read-only section `.initcall.9'
  /usr/bin/ld: warning: creating DT_TEXTREL in a PIE

This is understandable: If an initcall is emitted with a relocation,
marking its section as read-only is inconsistent.

Taking a step back, the actual problem is that we need to tell GCC that
our RO_DATA_SECTION is not executable. Therefore, let's insert it before
.data (which is rw^x) instead of .rodata, which is (rx).

Fixes: c11590f5643b ("sandbox: lds: fix "ELF has a LOAD segment with RWX permissions" warning")
Reported-by: Abdelrahman Youssef <abdelrahmanyossef12@gmail.com>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20241014113540.2817987-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>