dma: give dma_alloc_coherent a device parameter
dma_alloc_coherent takes a device parameter in Linux, which allows it to
device-specific non-1:1 mapping as well as different coherent DMA mask
and coherency settings.

To enable handling these quirks in future, let's add the device
parameter and set it to DMA_DEVICE_BROKEN for now.

This should introduce no functional change.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20241014125843.2064532-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>