test: enable JSON web token support in full.config
full.config is meant to be a snippet that enables as much as possible to
increase coverage beyond what's in the defconfig, e.g. to enable
selftests for a wider range of functionality.

We have no configs enabling CONFIG_JWT, so this is currently not tested.
Enable it in full.config to increase test coverage.

Next step would be enabling this for CI. This will have to wait until a
bthread related failure for x86 is fixed, which will be tackled
separately.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20241015082445.1670430-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>