ARM: Layerscape: TQMLS1046a: add support for 8GiB variant
The tqmls1046a also has a 8GiB DDR variant. This patch adds support for
this variant as an additional image with suffix _8g in the name.

The 8GiB boards use two different types of DDR chips which can be
distinguished by reading the EEPROM. There is one regular version which
has the string "TQMLS1046A-P2.0201" is the EEPROM. The other one is
possibly only used for Arkona and has the string "TQMLS1046A-CA.0202"
in the EEPROM. There might be other possible strings we are currently
not aware of, we fall back to the P2 variant in this case.

Link: https://lore.barebox.org/20241022-arm-layerscape-tfa-v2-4-69ac70ceb0ff@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>