ARM: Layerscape: remove register arguments
The various Layerscape xload_start_image functions all take r0, r1 and
r2 arguments. These are all unused in the code, so remove them.

The ls1021aiot board passes the memory size in r0 in one case, but this
is never used, so it's safe to remove that as well.

Link: https://lore.barebox.org/20241022-arm-layerscape-tfa-v2-6-69ac70ceb0ff@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>