ARM: Layerscape: LS1046a-rdb: Switch to TF-A support
Now that we can start TF-A on LS1046a based boards switch the LS1046ardb
over to use this rather than deprecated PPA.

Unlike the PPA the TF-A needs information about the amount of SDRAM
installed on the board. The LS1046a-rdb has a DDR4 DIMM connector, so
the SDRAM configuration must be runtime detected. The board support has
8GiB hardcoded in other places, so hardcode 8GiB for TF-A as well.

Link: https://lore.barebox.org/20241022-arm-layerscape-tfa-v2-8-69ac70ceb0ff@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>