commands: mmc_extcsd: split background operations
The current implementation of the  `mmc_extcsd -b` subcommand enables
manual background operations and automatic background operations if
available, which is unexpected and can be problematic:

With the eMMC JEDEC Standard 4.41 eMMCs gained support for decoupling
internal maintenance operations from foreground operations like read and
write commands. This is feature is called background operations. This
allows the host to manually trigger these maintenance operations e.g.
during idle times.

The advantages are reduced latencies during regular reads and writes. If
the host doesn't schedule these maintenance operations the standard
suggest[1] that they will be executed during foreground operations once
the maintenance level `critical` is reached. Thus enabling manual
background operations is a safe opt-in which resorts to maintenance
during foreground operations, even if the host isn't aware of the
enablement and doesn't start them at all.

With JEDEC standard 5.1 eMMCs gained support for schedueling these
background operations autonomously at their own will. This *not* a safe
opt-in as it needs support from the host:

> When AUTO_EN bit is set, the host should keep the device power active.

As there is no register to query the current status of a background
operation, the safe way seems to issue a power off notification and wait
for completion. This might not be implemented by the host.

This commit refactors the `mmc_extcsd -b` subcommand to only set the
`MANUAL_EN` bit, which is a safe default.

To still allow the enablement of  automatic background operations the
new `mmc_extcsd -B` subcommand is introduced which sets `MANUAL_EN` and
`AUTO_EN` bits.

[1]: See JEDEC Standard No. 84-B51, Page 93, 6.6.25 Background
Operations

Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>
Link: https://lore.kernel.org/r/20241023-feature-emmc-background-ops-v1-1-535161eeefec@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>