sandbox: sdl: Fix unresponsive SDL GUI by handling SDL_QUIT events
The SDL GUI in the scanout function can appear unresponsive to the
operating system, as it lacks an event handler. This patch adds a
new function, handle_sdl_events(), which processes SDL events to
ensure the window can respond to close requests.

This addresses issues where Linux might think the SDL window is
unresponsive, allowing it to handle SDL_QUIT events properly and
initiate shutdown.

Signed-off-by: Alaa Ali <alaanourali@gmail.com>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20241027081753.256921-1-alaanourali@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>