partitions: efi: fix overflow issues while allocating gpt entries
while parsting the GPT header in alloc_read_gpt_entries() the number
of partitions can be large that multiplying it with the size of a single
partition overflows 32-bit multiplication.

we already enforce a MAX_PARTITION limit of 128 partitions in efi_partition(),
so allowing any bigger value in alloc_read_gpt_entries() would fail,
even if we fix the overflow.

Therefore, we can enforce the limit strictly and treat any overflow as
a failing condition.

Signed-off-by: Abdelrahman Youssef <abdelrahmanyossef12@gmail.com>
Link: https://lore.kernel.org/r/20241031124854.625174-1-abdelrahmanyossef12@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>